+++ src/core/runtime/windows/arena.cs
@@ -208,18 +208,18 @@ internal static uintptr userArenaChunkPages => /* userArenaChunkBytes / pageSize
 internal static uintptr userArenaChunkMaxAllocBytes => /* userArenaChunkBytes / 4 */ 1048576;
 
 /* [GoInit] runtime bootstrap init - not run; .NET is the runtime */ internal static void init() {
-    if (userArenaChunkPages * (uintptr)pageSize != userArenaChunkBytes) {
+    if ((uintptr)(userArenaChunkPages * (uintptr)pageSize) != userArenaChunkBytes) {
         @throw("user arena chunk size is not a multiple of the page size"u8);
     }
     if (userArenaChunkBytes % physPageSize != 0) {
         @throw("user arena chunk size is not a multiple of the physical page size"u8);
     }
     if (userArenaChunkBytes < heapArenaBytes){
-        if ((uintptr)heapArenaBytes % userArenaChunkBytes != 0) {
+        if ((uintptr)((uintptr)heapArenaBytes % userArenaChunkBytes) != 0) {
             @throw("user arena chunk size is smaller than a heap arena, but doesn't divide it"u8);
         }
     } else {
-        if (userArenaChunkBytes % (uintptr)heapArenaBytes != 0) {
+        if ((uintptr)(userArenaChunkBytes % (uintptr)heapArenaBytes) != 0) {
             @throw("user arena chunks size is larger than a heap arena, but not a multiple"u8);
         }
     }
